partway through setting up basic ros1 node
